requirements.txt
playwright>=1.35.0
requests>=2.28.0
beautifulsoup4>=4.11.0
pandas>=2.1.0
python-dotenv>=0.19.0
lxml>=4.8.0
openpyxl>=3.0.0
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.0.0
pymongo>=4.6.0
dnspython>=2.4.0